+++ docs/CHANGELOG.md
@@ -6,10 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
-### Changed
 - Add the 3840x1200 resolution.
+- [Issue 125](https://github.com/aza547/wow-recorder/issues/125) - Push to talk for microphone recording.
 
+### Changed
 ### Fixed
+- [Issue 443](https://github.com/aza547/wow-recorder/issues/443) - Fix to Ctrl+R breaking the app.
 
 ## [3.18.1] - 2023-09-03
 ### Fixed